Fixed bugs
- Fix Velocity plugin startup
- Fix error with permissions on paper
- Fix... paper in velocity and velocity in paper
