+++ pom.xml
@@ -6,7 +6,7 @@
 
     <groupId>net.dreamerzero.titleannouncer</groupId>
     <artifactId>TitleAnnouncer</artifactId>
-    <version>1.17.0</version>
+    <version>2.0.0</version>
     <packaging>jar</packaging>
 
     <properties>
