+++ src/main/java/net/dreamerzero/titleannouncer/common/utils/BossBarUtils.java
@@ -1,111 +1,9 @@
 package net.dreamerzero.titleannouncer.common.utils;
 
-import java.util.concurrent.TimeUnit;
-
-import com.velocitypowered.api.proxy.ProxyServer;
-import com.velocitypowered.api.scheduler.ScheduledTask;
-
-import org.bukkit.scheduler.BukkitRunnable;
-
-import net.dreamerzero.titleannouncer.paper.Announcer;
 import net.kyori.adventure.audience.Audience;
 import net.kyori.adventure.bossbar.BossBar;
-import net.kyori.adventure.text.Component;
 
 public class BossBarUtils {
-    private static float value = 1f;
-
-    /**
-     * It will send a bossbar to the specified audience,
-     * with the specified characteristics showing an
-     * emptying animation according to the specified
-     * time interval.
-     * @param audience
-     * @param time
-     * @param content
-     * @param color
-     * @param type
-     */
-    public static void sendBukkitBossBar (
-        final Audience audience,
-        final float time,
-        final Component content,
-        final BossBar.Color color,
-        final BossBar.Overlay type) {
-
-        float finalTime = 0.1f/time;
-
-        BossBar bar = BossBar.bossBar(
-            content,
-            1,
-            color,
-            type);
-
-        audience.showBossBar(bar);
-        final float toReduce = finalTime;
-        value = 1f;
-
-        new BukkitRunnable() {
-            @Override
-            public void run() {
-                value -= toReduce;
-                if (value <= 0.02) {
-                    audience.hideBossBar(bar);
-                    cancel();
-                }
-                try {
-                    bar.progress(value);
-                } catch (IllegalArgumentException e) {
-                    cancel();
-                }
-            }
-        }.runTaskTimerAsynchronously(Announcer.getInstance(), 20, 2);
-    }
-
-    public static void sendVelocityBossbar(
-        final Audience audience,
-        final float time,
-        final Component content,
-        final BossBar.Color color,
-        final BossBar.Overlay type){
-
-        float finalTime = 0.1f/time;
-
-        BossBar bar = BossBar.bossBar(
-            content,
-            1,
-            color,
-            type);
-
-        audience.showBossBar(bar);
-        final float toReduce = finalTime;
-        value = 1f;
-
-        ProxyServer server = net.dreamerzero.titleannouncer.velocity.Announcer.getProxyServer();
-
-        server.getScheduler()
-        .buildTask(net.dreamerzero.titleannouncer.velocity.Announcer.getVInstance(), () -> {
-            value -= toReduce;
-                if (value <= 0.02) {
-                    audience.hideBossBar(bar);
-                    return;
-                }
-                try {
-                    bar.progress(value);
-                } catch (IllegalArgumentException e) {
-                    return;
-                }
-        })
-        .repeat(2000L, TimeUnit.MILLISECONDS)
-        .schedule();
-        //scheduler.cancel();
-
-    }
-
-    public static void cancelTask(ScheduledTask task){
-        task.cancel();
-    }
-
     /**
      * Based on the argument given in the command, 
      * it will return the color of the specified bossbar.
