+++ src/main/java/net/dreamerzero/titleannouncer/paper/commands/title/AnnouncerTitleCommand.java
@@ -24,7 +24,7 @@ public class AnnouncerTitleCommand implements CommandExecutor {
     // Command
     public boolean onCommand(CommandSender sender, Command cmd, String label, String[] args) {
         // Permission Check
-        if (sender.permissionValue("announcer.title.global") != TriState.TRUE) {
+        if (sender.permissionValue("titleannouncer.title.global") != TriState.TRUE) {
             ConfigUtils.sendNoMainPermission(sender);
             return true;
         }
