+++ src/main/java/net/dreamerzero/titleannouncer/velocity/commands/bossbar/ServerBossbarCommand.java
@@ -14,6 +14,7 @@ import net.dreamerzero.titleannouncer.common.utils.GeneralUtils;
 import net.dreamerzero.titleannouncer.common.utils.MiniMessageUtil;
 import net.dreamerzero.titleannouncer.common.utils.PlaceholderUtil;
 import net.dreamerzero.titleannouncer.common.utils.SoundUtil;
+import net.dreamerzero.titleannouncer.velocity.utils.VelocityBossbar;
 import net.kyori.adventure.bossbar.BossBar;
 
 public class ServerBossbarCommand implements SimpleCommand {
@@ -56,7 +57,7 @@ public class ServerBossbarCommand implements SimpleCommand {
         // Concatenate the arguments provided by the command sent.
         String bossbartext = GeneralUtils.getCommandString(args, 5);
 
-        BossBarUtils.sendVelocityBossbar(
+        VelocityBossbar.sendVelocityBossbar(
             serverObjetive,
             time,
             MiniMessageUtil.parse(
